requirements.txt
allure-python-commons
requests
pytest
selenium
